Session 27: Fix DXVK header conflicts (MEMORYSTATUS, IUnknown)
Major achievements:
- Resolved MEMORYSTATUS type conflict between GeneralsX and DXVK
- Resolved IUnknown interface conflict between com_compat and DXVK
- Added conditional compilation guards (_MEMORYSTATUS_DEFINED, _IUNKNOWN_DEFINED)
- Build progress: 334/905 files (36.9%) - NEW RECORD!

Technical changes:
- Patched DXVK windows_base.h and unknwn.h with #ifndef guards
- Updated memory_compat.h with full 8-field MEMORYSTATUS struct
- Updated com_compat.h with DXVK coexistence guard
- Renamed windows.h to windows_wrapper.h (avoid include collision)

Files modified:
- GeneralsMD/Code/CompatLib/Include/memory_compat.h
- GeneralsMD/Code/CompatLib/Include/com_compat.h
- GeneralsMD/Code/CompatLib/Include/windows.h
- GeneralsMD/Code/GameEngine/Include/Precompiled/PreRTS.h
- docs/DEV_BLOG/2026-02-DIARY.md (Session 27 entry)

Plus DXVK build tree patches (documented in dev blog)

TheSuperHackers @build BenderAI 12/02/2026

+++ GeneralsMD/Code/GameEngine/Source/Common/Thing/ThingTemplate.cpp
@@ -506,7 +506,8 @@ void ThingTemplate::parseModuleName(INI* ini, void *instance, void* store, const
 {
 	ThingTemplate* self = (ThingTemplate*)instance;
 	ModuleInfo* mi = (ModuleInfo*)store;
-	ModuleType type = (ModuleType)(UnsignedInt)userData;
+	// TheSuperHackers @build fighter19 11/02/2026 Bender - Use uintptr_t for pointer-to-int cast (64-bit compat)
+	ModuleType type = (ModuleType)(UnsignedInt)(uintptr_t)userData;
 	const char* token = ini->getNextToken();
 	AsciiString tokenStr = token;
 
@@ -613,7 +614,8 @@ void ThingTemplate::parseModuleName(INI* ini, void *instance, void* store, const
 //-------------------------------------------------------------------------------------------------
 void ThingTemplate::parseIntList(INI* ini, void *instance, void* store, const void* userData)
 {
-	Int numberEntries = (Int)userData;
+	// TheSuperHackers @build fighter19 11/02/2026 Bender - Use uintptr_t for pointer-to-int cast (64-bit compat)
+	Int numberEntries = (Int)(uintptr_t)userData;
 	Int *intList = (Int*)store;
 
 	for( Int intIndex = 0; intIndex < numberEntries; intIndex ++ )
